+++ examples/libadwaita/tab_factory.rs
@@ -1,7 +1,7 @@
 use gtk::prelude::{BoxExt, ButtonExt, GtkWindowExt, OrientableExt};
 use relm4::{
     adw,
-    factory::{DynamicIndex, FactoryComponent, FactoryComponentSender, FactoryVecDeque},
+    factory::{DynamicIndex, FactoryComponent, FactorySender, FactoryVecDeque},
     gtk, ComponentParts, ComponentSender, RelmApp, RelmWidgetExt, SimpleComponent,
 };
 
@@ -94,15 +94,11 @@ impl FactoryComponent for Counter {
         })
     }
 
-    fn init_model(
-        value: Self::Init,
-        _index: &DynamicIndex,
-        _sender: FactoryComponentSender<Self>,
-    ) -> Self {
+    fn init_model(value: Self::Init, _index: &DynamicIndex, _sender: FactorySender<Self>) -> Self {
         Self { value }
     }
 
-    fn update(&mut self, msg: Self::Input, _sender: FactoryComponentSender<Self>) {
+    fn update(&mut self, msg: Self::Input, _sender: FactorySender<Self>) {
         match msg {
             CounterMsg::Increment => {
                 self.value = self.value.wrapping_add(1);