+++ relm4/src/factory/mod.rs
@@ -18,6 +18,6 @@ use data_guard::DataGuard;
 pub use r#async::{AsyncFactoryComponent, AsyncFactoryVecDeque, AsyncFactoryVecDequeGuard};
 pub use sync::{FactoryComponent, FactoryVecDeque, FactoryVecDequeGuard};
 
-pub use crate::sender::{AsyncFactoryComponentSender, FactoryComponentSender};
+pub use crate::channel::{AsyncFactorySender, FactorySender};
 pub use dynamic_index::DynamicIndex;
 pub use widgets::traits::*;