+++ CHANGES.md
@@ -7,7 +7,11 @@
 + core: Add asynchronous components including macro support
 + core: Add asynchronous factories including macro support
 + core: Temporary widget initialization for async components and factories
++ core: Add `LoadingWidgets` to help with temporary loading widgets in async factories and components
 + core: Add `Reducer` as message based alternative to `SharedState`
++ core: Synchronous API for commands
++ core: Add `try_send` method for fallible sending of messages
++ core: Remove async-broadcast dependency
 + core: Runtimes of `Component`s and `AsyncComponents` can now be detached for a static lifetime
 + core: Add `ComponentStream` as alternative to `Controller` that implements `Stream` for async message handling
 + core: Add `gnome_42` and `gnome_43` feature flags
@@ -17,6 +21,10 @@
 
 ### Changed
 
++ core: Rename `FactoryComponentSender` to `FactorySender` and `AsyncFactoryComponentSender` to `AsyncFactorySender`
++ core: The sender API now supports proper error handling
++ core: Add `gnome_42` and `gnome_43` feature flags
++ core: Pass `Root` during `update` and `update_cmd` for `Component` and `AsyncComponent`
 + core: Rename `OnDestroy` to `RelmObjectExt`
 + core: Remove `EmptyRoot` in favor of the unit type
 + macros: Allow using methods calls as widget initializers in the `view` macro